+++ CHANGELOG.md
@@ -1,5 +1,16 @@
 ### Changelog
 
+##### v1.8.0
++ [PSR-2 coding style guide](http://www.php-fig.org/psr/psr-2/) adopted 
++ Updated `InsertStatement` class with:
+  - Added `columns()` method
++ Updated `UpdateStatement` class with:
+  - Added `set()` method
++ Updated `StatementContainer` class with:
+  - Added `$table` argument in `delete()` method
++ Updated `WhereClause` class with:
+  - Fixed `orWhereLike()`
+
 ##### v1.7.2
 + Updated `SelectStatement` class with:
   - Minor fix `select()` method (working fix)